// src/openapi/models/OrganizationDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface OrganizationDto
 */
export interface OrganizationDto {
    /**
     * 
     * @type {string}
     * @memberof OrganizationDto
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof OrganizationDto
     */
    leader?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof OrganizationDto
     */
    members?: Set<string>;
    /**
     * 
     * @type {string}
     * @memberof OrganizationDto
     */
    parentOrganization?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof OrganizationDto
     */
    subordinateOrganizations?: Set<string>;
    /**
     * 
     * @type {string}
     * @memberof OrganizationDto
     */
    name?: string;
}

export function OrganizationDtoFromJSON(json: any): OrganizationDto {
    return OrganizationDtoFromJSONTyped(json, false);
}

export function OrganizationDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrganizationDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'leader': !exists(json, 'leader') ? undefined : json['leader'],
        'members': !exists(json, 'members') ? undefined : json['members'],
        'parentOrganization': !exists(json, 'parentOrganization') ? undefined : json['parentOrganization'],
        'subordinateOrganizations': !exists(json, 'subordinateOrganizations') ? undefined : json['subordinateOrganizations'],
        'name': !exists(json, 'name') ? undefined : json['name'],
    };
}

export function OrganizationDtoToJSON(value?: OrganizationDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'leader': value.leader,
        'members': value.members,
        'parentOrganization': value.parentOrganization,
        'subordinateOrganizations': value.subordinateOrganizations,
        'name': value.name,
    };
}